+++ packages/developer-portal/src/components/pages/webhooks/webhooks-logs.tsx
@@ -1,0 +1,5 @@
+import React, { FC } from 'react'
+
+export const WebhooksLogs: FC = () => {
+  return <div>Logs Webhook</div>
+}
