+++ packages/developer-portal/src/components/pages/webhooks/webhooks-manage.tsx
@@ -1,0 +1,5 @@
+import React, { FC } from 'react'
+
+export const WebhooksManage: FC = () => {
+  return <div>Manage Webhook</div>
+}
