+++ packages/developer-portal/src/components/pages/webhooks/webhooks-new.tsx
@@ -1,0 +1,5 @@
+import React, { FC } from 'react'
+
+export const WebhooksNew: FC = () => {
+  return <div>New Webhook</div>
+}
