+++ packages/developer-portal/src/constants/pages.ts
@@ -7,6 +7,9 @@ export const ELEMENTS_V3_PAGES = [
   Routes.SWAGGER,
   Routes.API_DOCS,
   Routes.APPS,
-  Routes.WEBHOOKS,
   Routes.ANALYTICS_SCHEMA_DOCS,
+  Routes.WEBHOOKS_ABOUT,
+  Routes.WEBHOOKS_LOGS,
+  Routes.WEBHOOKS_MANAGE,
+  Routes.WEBHOOKS_NEW,
 ]
