+++ packages/developer-portal/src/constants/routes.ts
@@ -12,7 +12,10 @@ const Routes = {
   ANALYTICS: '/analytics',
   ANALYTICS_TAB: '/analytics/:activeTab?',
   RESET_PASSWORD: '/reset-password',
-  WEBHOOKS: '/webhooks',
+  WEBHOOKS_ABOUT: '/webhooks/about',
+  WEBHOOKS_NEW: '/webhooks/new',
+  WEBHOOKS_MANAGE: '/webhooks/manage',
+  WEBHOOKS_LOGS: '/webhooks/logs',
   SETTINGS: '/settings/',
   SETTINGS_PROFILE_TAB: '/settings/profile',
   SETTINGS_ORGANISATION_TAB: '/settings/organisation',
