+++ packages/elements/src/storybook/changelog.stories.mdx
@@ -14,6 +14,10 @@ We will publish release version history and changes here. Where possible, we wil
 
 Beta versions should be relatively stable but subject to occssional breaking changes.
 
+### **3.4.0 - 10/09/21**
+
+- `Tabs` full screen now aligns left
+
 ### **3.4.0 - 27/08/21**
 
 - Adds new `Tabs` component
